StructuralAnalysis Linear and NonLinear static (#153)
